Fixed comment text layer size
The text layer holding comment thread text would only decrese in size
when calling text_layer_set_text(). Instead I now delete the layer and
rebuild it every time the window loads (regardless if it's the first
comment in a thread or the window has already been initialized). This
appears to fix it.